Make new utilities module
